search/requirements/base.txt
alembic==1.12.1
aiofiles==23.2.1
aio_pika==9.3.0
asyncpg==0.28.0
fastapi==0.104.1
fastapi_mail==1.4.1
fastapi_pagination==0.12.12
fastapi-sqlalchemy-toolkit==0.7.0.1
fastapi_users[sqlalchemy,oauth]==12.1.2
pillow==10.2.0
pydantic==2.4.2
pydantic-settings==2.0.3
setuptools==68.2.2
uvicorn==0.23.2
sqlalchemy==2.0.22
sqlakeyset==2.0.1695177552
tenacity==8.2.3
pytz==2024.1
websockets==12.0
broadcaster==0.2.0
asyncio_redis==0.16.0

#triton
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
Brotli==1.1.0
certifi==2024.2.2
click==8.1.7
frozenlist==1.4.1
gevent==24.2.1
geventhttpclient==2.0.2
greenlet==3.0.3
idna==3.6
multidict==6.0.5
numpy==1.26.4
opencv-python==4.9.0.80
python-rapidjson==1.16
six==1.16.0
tritonclient==2.44.0
urllib3==2.2.1
yarl==1.9.4
youtokentome @ git+https://github.com/gburlet/YouTokenToMe.git@e0fec1f6e52fc75a841ed3fab8f583448b1deedf
zope.event==5.0
zope.interface==6.2
pgvector==0.2.5
